refactor(tables): rename dual-framework files to generic names
The `data/tables/crr_haircuts.py` and `data/tables/crr_firb_lgd.py` files
both held CRR and Basel 3.1 tables side-by-side, so the `crr_` prefix
misled readers. Rename both to `haircuts.py` and `firb_lgd.py` to match
the generic-name convention already used for other dual-framework
tables (`crm_supervisory.py`, `eu_sovereign.py`, `output_floor.py`).

`data/tables/b31_firb_lgd.py` was a thin re-export of the Basel 3.1
LGD dict that physically lived in `crr_firb_lgd.py`. Merge its contents
(B31 scalar constants, `lookup_b31_firb_lgd`, `get_b31_firb_lgd_table`,
`get_b31_vs_crr_lgd_comparison`) into the new `firb_lgd.py` so there is
a single source of truth, then delete it.

Module docstrings updated to reflect dual-framework scope. Imports
rewritten across `engine/`, tests, and docs. `data/tables/__init__.py`
re-exports preserved; no public API changes and no regulatory values
changed.

Verified: ruff, ty, arch_check, and unit/contracts test suites all pass.
