+++ src/rwa_calc/data/tables/__init__.py
@@ -8,11 +8,10 @@ as onshored into UK law, plus Basel 3.1 (PRA PS1/26) tables.
 Modules:
     crr_risk_weights: SA risk weights by exposure class and CQS
     b31_risk_weights: Basel 3.1 LTV-band SA risk weights for real estate
-    crr_haircuts: CRM supervisory haircuts
+    haircuts: CRM supervisory haircuts (CRR Art. 224 and PRA PS1/26 Art. 224)
     crr_slotting: CRR specialised lending slotting risk weights
     b31_slotting: Basel 3.1 specialised lending slotting risk weights
-    crr_firb_lgd: F-IRB supervisory LGD values (CRR)
-    b31_firb_lgd: F-IRB supervisory LGD values (Basel 3.1)
+    firb_lgd: F-IRB supervisory LGD values (CRR Art. 161 and PRA PS1/26 Art. 161)
     crr_equity_rw: CRR equity risk weights (Art. 133 SA, Art. 155 IRB Simple)
     b31_equity_rw: Basel 3.1 equity risk weights (PRA PS1/26 Art. 133)
 """
@@ -22,20 +21,6 @@ from .b31_equity_rw import (
     get_b31_equity_risk_weights,
     get_b31_equity_rw_table,
     lookup_b31_equity_rw,
-)
-from .b31_firb_lgd import (
-    B31_FIRB_LGD_COMMERCIAL_RE,
-    B31_FIRB_LGD_COVERED_BOND,
-    B31_FIRB_LGD_FINANCIAL_COLLATERAL,
-    B31_FIRB_LGD_OTHER_PHYSICAL,
-    B31_FIRB_LGD_RECEIVABLES,
-    B31_FIRB_LGD_RESIDENTIAL_RE,
-    B31_FIRB_LGD_SUBORDINATED,
-    B31_FIRB_LGD_UNSECURED_SENIOR,
-    B31_FIRB_LGD_UNSECURED_SENIOR_FSE,
-    get_b31_firb_lgd_table,
-    get_b31_vs_crr_lgd_comparison,
-    lookup_b31_firb_lgd,
 )
 from .b31_risk_weights import (
     B31_ADC_PRESOLD_RISK_WEIGHT,
@@ -66,18 +51,6 @@ from .crr_equity_rw import (
     get_equity_rw_table,
     lookup_equity_rw,
 )
-from .crr_firb_lgd import (
-    BASEL31_FIRB_SUPERVISORY_LGD,
-    FIRB_SUPERVISORY_LGD,
-    get_firb_lgd_table,
-    get_firb_lgd_table_for_framework,
-)
-from .crr_haircuts import (
-    BASEL31_COLLATERAL_HAIRCUTS,
-    COLLATERAL_HAIRCUTS,
-    FX_HAIRCUT,
-    get_haircut_table,
-)
 from .crr_risk_weights import (
     CENTRAL_GOVT_CENTRAL_BANK_RISK_WEIGHTS,
     COMMERCIAL_RE_PARAMS,
@@ -100,6 +73,30 @@ from .eu_sovereign import (
     EU_COUNTRY_DOMESTIC_CURRENCY,
     EU_MEMBER_STATES,
     build_eu_domestic_currency_expr,
+)
+from .firb_lgd import (
+    B31_FIRB_LGD_COMMERCIAL_RE,
+    B31_FIRB_LGD_COVERED_BOND,
+    B31_FIRB_LGD_FINANCIAL_COLLATERAL,
+    B31_FIRB_LGD_OTHER_PHYSICAL,
+    B31_FIRB_LGD_RECEIVABLES,
+    B31_FIRB_LGD_RESIDENTIAL_RE,
+    B31_FIRB_LGD_SUBORDINATED,
+    B31_FIRB_LGD_UNSECURED_SENIOR,
+    B31_FIRB_LGD_UNSECURED_SENIOR_FSE,
+    BASEL31_FIRB_SUPERVISORY_LGD,
+    FIRB_SUPERVISORY_LGD,
+    get_b31_firb_lgd_table,
+    get_b31_vs_crr_lgd_comparison,
+    get_firb_lgd_table,
+    get_firb_lgd_table_for_framework,
+    lookup_b31_firb_lgd,
+)
+from .haircuts import (
+    BASEL31_COLLATERAL_HAIRCUTS,
+    COLLATERAL_HAIRCUTS,
+    FX_HAIRCUT,
+    get_haircut_table,
 )
 
 __all__ = [
@@ -143,12 +140,12 @@ __all__ = [
     "B31_SLOTTING_RISK_WEIGHTS_PREOP",
     "B31_SLOTTING_RISK_WEIGHTS_HVCRE",
     "lookup_b31_slotting_rw",
-    # F-IRB LGD — CRR
+    # F-IRB LGD — dual framework
     "FIRB_SUPERVISORY_LGD",
     "BASEL31_FIRB_SUPERVISORY_LGD",
     "get_firb_lgd_table",
     "get_firb_lgd_table_for_framework",
-    # F-IRB LGD — Basel 3.1
+    # F-IRB LGD — Basel 3.1 scalar constants
     "B31_FIRB_LGD_UNSECURED_SENIOR",
     "B31_FIRB_LGD_UNSECURED_SENIOR_FSE",
     "B31_FIRB_LGD_SUBORDINATED",
